Brand Page, Details Page, Private Routing with Authentication:
This commit creates the 'Brand' page and the 'Details' page according to the brands displayed in the brand section of home in the form of cards. The 'Brand' page cards have product information and two buttons: 1. details, 2. update. The 'Details' page loads a details page of the particular product with relavent information about it. This commit also implements most of the required private routings. So without logging in, the users won't be able to access the private/protected routes/pages.