'Update-Product' Page, Update information in MongoDB:
This commit implements the 'Update-Product' page. The way to get to this page is through the update button in 'Brand-Products' page or the update button in the 'My-Cart' page. This page is private/protected, only logged in users can access this page. This commit also handles all the backend functionality as well, which are: sending the updated data to the server through PUT api, then updating the data in the database from the server. Once the updating is done, the user will be redirected to the 'Brand-Producst' page, where one of the update button exists.