ErrorPage Added:
This commit implements the basic ErrorPage, the error page has a button saying 'Go Back To Home', clicking on which will lead the user back to the home page. This ErrorPage will load whenever the user tries to load any url on the midnight-tales website that doesn't exist. The message on the ErrorPage is a simple '404 - Page Not Found'.