Added README.md File:
This commit adds a readme.md file in the root directory of the project for user's and developers. This readme.md file mentions over 5 features of the midnight tales website, displays a preview of the code and adds badges of the libraries and resources that were used in this project. This commit also includes a live link to the actual website that the users can click and navigate to.