DarkMode enhanced & Select Box added for adding brand names:
This commit implements the dark mode to work even after the user refreshes or moves to a different page by using the local storage of the browser to save that information. This commit also changes the input text field to a select box in the 'Add Product' page in order to not have the users make any mistake when adding the brand name because it is being used to get information from the database.